Refactor the file structure
