+++ classes/student_class.rb
