+++ classes/trimmer_decorator.rb
